+++ Controllers/Global.cs
@@ -10,6 +10,8 @@ public partial class Global : Node
     public static Global Instance;
     public static GameController.GameController GameController;
     public static EconomyController.EconomyController Economy { get; } = new EconomyController.EconomyController();
+    public static GameController GameController;
+    public static EconomyController Economy { get; } = new();
     public static Scenes.Player.Player Player;
     
     internal Logger Logger;
